requirements.txt
apache-airflow-providers-apache-spark==5.4.0
pyspark==3.5.2
